requirements.txt
Django~=4.0.2
django-colorfield
numpy
django-crontab
